Change the style and look of the page
Add color to the content section as well as the comments to
stand out more. Change spacing so content isn't on the border
of its section and isn't too close to other items.

Place the comment submission below the comments like most website
comment sections. Make comments into individual boxes instead
of bullet points in a list.

Change the image section from random images to a slideshow.
Standardize image sizes to keep the size of the section constant
other than space to grow for comments.