Add image submission to comments
Add servlet to create a url for the submission form to send data to
Blobstore for image storage.

Add function to DataServlet.java to retrieve an image from Blobstore
and store it in Datastore with the rest of the comment. Uses the
metadata of the image if available to correct image orientation
if the rotation is wrong. Stores the changed image in Google
Cloud Storage and saves the url to the image.